Raise exception if SSH is unconfigured
Previous behavior would run a command locally if the ssh proxy
credentials weren't configured. That leads to false positives, though,
where a call to _run() succeeds but not against the remote target.

Calls to _run() will now raise an Exception if the SSH credentials
aren't configured. Expect this to be refined as more use-cases are
tested.