Fix window_size overflow
It was possible for commands to return so much data via stdout that it
would cause Paramiko's ssh channel to hang:

> In some situations, receiving remote output larger than the current
> Transport or session’s window_size (e.g. that set by the
> default_window_size kwarg for Transport.__init__) will cause
> recv_exit_status to hang indefinitely if it is called prior to a
> sufficiently large Channel.recv (or if there are no threads calling
> Channel.recv in the background).

> In these cases, ensuring that recv_exit_status is called after
> Channel.recv (or, again, using threads) can avoid the hang.

One way to work around this was to pipe the output of verbose scripts to
/dev/null. It's still good practice to quiet output not necessary for
your command to run, because stdout is still transmitted over the ssh
channel.

This change implements a buffered reading of stdout to prevent the type
of hang described above

This also contains some refactoring to ensure methods called by an
action only run in an action context, and adds additional pydoc.